[FIX] When a file in the directory cannot be deleted, the folder cannot be
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@45628 b456876b-0849-0410-b77d-98878d47e9d5

+++ lib/core/Tiki/Composer/CleanVendors.php
@@ -80,7 +80,12 @@ class CleanVendors
 			}
 		}
 
-		if (is_writable($path)) {
+		$remaining = scandir($path);
+		$remaining = array_filter($remaining, function ($filename) {
+			return $file !== '.' && $file !== '..';
+		});
+
+		if (is_writable($path) && ! count($remaining)) {
 			rmdir($path);
 		}
 	}
